runtime/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mendicant-runtime"
version = "5.0.0"
description = "Mendicant Bias Agent Runtime — LangGraph agent with intelligence middleware"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Mendicant Bias Project" },
]
keywords = ["ai", "agents", "runtime", "langgraph", "mendicant"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "mendicant-core>=5.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langchain-openai>=0.3.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=3.0.0",
]
tokens = [
    "tiktoken>=0.8.0",
]
all = [
    "mendicant-runtime[embeddings,tokens]",
]

[tool.hatch.build.targets.wheel]
packages = ["src/mendicant_runtime"]

[tool.pytest.ini_options]
testpaths = ["tests"]
